agents/openhands_agent/requirements.txt
openhands-sdk>=1.10.0
openhands-tools>=1.10.0
openhands-workspace>=1.10.0
